mv aws_session_expiration_check to tests/conf
